requirements.txt
playwright>=1.44.0
beautifulsoup4>=4.12.3
requests>=2.31.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
flask>=3.0.3
flask-cors>=4.0.1
